resources/observability/python-app/requirements.txt
flask
azure-monitor-opentelemetry
opentelemetry-exporter-otlp
